+++ src/main/java/io/github/mortuusars/monobank/content/monobank/MonobankBlock.java
@@ -219,6 +219,11 @@ public class MonobankBlock extends Block implements EntityBlock {
 
         boolean isLocked = monobankEntity.getLock().isLocked();
 
+        if (isLocked && !Configuration.COMBINATION_ENABLED.get()) {
+            monobankEntity.startUnlocking();
+            return InteractionResult.SUCCESS;
+        }
+
         if (Configuration.OWNER_CAN_UNLOCK_WITHOUT_COMBINATION.get() && monobankEntity.getOwner().isOwnedBy(player)) { // Lock/Unlock to the heart's content
             boolean shouldBeLocked = !isLocked;