+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Monobanks are no longer carryable with `Carry On` mod.
 - Fixed 'Break In Attempted' or 'Break In Successful' warnings being resettable by a player that does not own the Monobank.
 - Break In warnings no longer show for a player that does not own the Monobank.
+- Added several config options related to combination.
 
 ### 1.1.4 - 2024-02-07
 - Fixed Monobanks in most villages not having a loot [1.20.1]